refactoring foulder adminPage
