+++ src/app/components/common/table/tableAdmin/index.js
