+++ src/app/components/common/table/tableAdmin/table.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import AdminHeader from "./adminHeader";
-import AdminBody from "./adminBody";
+import AdminHeader from "./tableHeader";
+import AdminBody from "./tableBody";
 
-const AdminTable = ({ onSort, selectedSort, columns, data, children }) => {
+const Table = ({ onSort, selectedSort, columns, data, children }) => {
     return (
         <table className="table table-borderless mx-2">
             {children || (
@@ -17,7 +17,7 @@ const AdminTable = ({ onSort, selectedSort, columns, data, children }) => {
     );
 };
 
-AdminTable.propTypes = {
+Table.propTypes = {
     onSort: PropTypes.func,
     selectedSort: PropTypes.object,
     columns: PropTypes.object,
@@ -25,4 +25,4 @@ AdminTable.propTypes = {
     children: PropTypes.array
 };
 
-export default AdminTable;
+export default Table;
