+++ src/app/components/common/table/tableProducts/table.jsx
