+++ src/app/components/page/adminPage/index.js
@@ -1,0 +1,2 @@
+import AdminPage from "./adminPage";
+export default AdminPage;
